% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCounts.R
\name{featureCounts}
\alias{featureCounts}
\title{featureCounts}
\usage{
featureCounts(gtf, bam, outputDir, feature = "exon", cores = 10)
}
\arguments{
\item{gtf}{}

\item{bam}{}

\item{outputDir}{}

\item{feature}{}

\item{cores}{}
}
\description{
featureCounts
}
\examples{

}
