% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrectedIsoform.R
\name{generateCorrectedIsoform}
\alias{generateCorrectedIsoform}
\title{Title}
\usage{
generateCorrectedIsoform(LRjunc, SRjunc, LRtag, LRread, ts, td,
  matchedLS = NULL)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{LRtag}{}

\item{LRread}{}

\item{ts}{}

\item{td}{}

\item{matchedLS}{}
}
\description{
Title
}
