% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrectedIsoform0.R
\name{generateCorrectedIsoform0}
\alias{generateCorrectedIsoform0}
\title{generateCorrectedIsoform0}
\usage{
generateCorrectedIsoform0(LRjunc, SRjunc, LRtag, LRread, ts, td,
  matchedLS = NULL, fuzzyMatch = 100)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{LRtag}{}

\item{LRread}{}

\item{ts}{}

\item{td}{}

\item{matchedLS}{}

\item{fuzzyMatch}{}
}
\description{
generateCorrectedIsoform0
}
