% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrectedJunc.R
\name{generateCorrectedJunc}
\alias{generateCorrectedJunc}
\title{generateCorrectedJunc}
\usage{
generateCorrectedJunc(LRjunc, SRjunc, LRread, matchedLS, ts, td)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{LRread}{}

\item{matchedLS}{}

\item{ts}{}

\item{td}{}
}
\description{
generateCorrectedJunc
}
