% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeq.R
\name{generateSeq}
\alias{generateSeq}
\title{generateSeq}
\usage{
generateSeq(fasta, Junc, chrom, s, e)
}
\arguments{
\item{fasta}{}

\item{Junc}{}

\item{chrom}{}

\item{s}{}

\item{e}{}
}
\description{
generateSeq
}
