% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeq2.R
\name{generateSeq2}
\alias{generateSeq2}
\title{generateSeq2}
\usage{
generateSeq2(ref, isoform, exp, chrom, s, e)
}
\arguments{
\item{ref}{}

\item{chrom}{}

\item{s}{}

\item{e}{}
}
\description{
generateSeq2 generates the transcript sequence
}
\examples{


}
