% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJunc.R
\name{getJunc}
\alias{getJunc}
\title{getJunc}
\usage{
getJunc(alignments, outputDir, chrom = NULL, s, e)
}
\arguments{
\item{alignments}{a character value, indicating the alignment file, .sam for .bam}

\item{outputDir}{a character value, indicating the directory whether to save the results}

\item{chrom}{a vector of character values, incidating the chromosomes of insterest}

\item{s}{an integer value indicating the start site}

\item{e}{an integer value indicating the end site}
}
\description{
getJunc gets the unique junction sites and their numbers from the alignment files
}
