% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJuncBySam.R
\name{getJuncBySam}
\alias{getJuncBySam}
\title{getJuncBySam}
\usage{
getJuncBySam(alignments, outputDir, chrom = NULL, s = 0, e = Inf)
}
\arguments{
\item{alignments}{a character value, indicating the alignment file, .sam for .bam}

\item{outputDir}{a character value, indicating the directory whether to save the results}

\item{chrom}{a vector of character values, incidating the chromosomes of insterest}

\item{s}{an integer value indicating the start site}

\item{e}{an integer value indicating the end site}
}
\value{
a list of data frame representing the splicing junction sites for 
each chromosome
}
\description{
getJuncBySam gets the unique junction sites and their numbers from the alignment files
}
