% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJuncBySJout.R
\name{getJuncBySJout}
\alias{getJuncBySJout}
\title{getJuncBySJout}
\usage{
getJuncBySJout(SJout = "SJ.out.tab", outputDir, chrom = NULL, s = 0,
  e = Inf)
}
\arguments{
\item{SJout}{a character value, indicating the SJ.out.tab file}

\item{outputDir}{a character value, indicating the directory whether to save the results}

\item{chrom}{a vector of character values, incidating the chromosomes of insterest}

\item{s}{an integer value indicating the start site}

\item{e}{an integer value indicating the end site}
}
\description{
getJuncBySjout gets the unique junction sites and their numbers from the alignment files
}
\examples{

}
