% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRinfo.R
\name{getLRinfo}
\alias{getLRinfo}
\title{getLRinfo
get the information for each long read}
\usage{
getLRinfo(LRread, chrom = NULL, s = 0, e = Inf)
}
\arguments{
\item{LRread}{a list of data frame, indicating the information of long read,
comprising read id, chr, strand, start, end and junction sites}

\item{chrom}{a character value, incidating the chromosome of interest}

\item{s}{an integer value indicating the start site of gene of interest. This
is useful for targeted sequencing}

\item{e}{an integer value indicating the end site of gene of interest. This
is useful for targeted sequencing}
}
\description{
getLRinfo
get the information for each long read
}
\examples{


}
