% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadByCigar.R
\name{getReadByCigar}
\alias{getReadByCigar}
\title{getReadByCigar}
\usage{
getReadByCigar(alignments, outputDir)
}
\arguments{
\item{alignments}{}

\item{outputDir}{}

\item{intronMinSize}{}
}
\description{
getReadByCigar
}
\examples{

}
