% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReadByJI.R
\name{getReadByJI}
\alias{getReadByJI}
\title{getReadByJI
get the long read information from JI tag in sam file}
\usage{
getReadByJI(alignments, outputDir)
}
\arguments{
\item{alignments}{character value indicating the location of alignment files,
such as .bam or .sam files}

\item{outputDir}{character value indicating the output directory for saving 
results}
}
\value{
a list of data frame, indicating the information of long read,
comprising id, chr, strand, start, end, junc
}
\description{
getReadByJI
get the long read information from JI tag in sam file
}
