% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridSearch}
\alias{gridSearch}
\title{gridSearch}
\usage{
gridSearch(LRjunc, SRjunc, matchedLS, thresSR = c(1:30), thresDis = c(1:20))
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{matchedLS}{}

\item{thresSR}{}

\item{thresDis}{}
}
\description{
gridSearch
}
