% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch0.R
\name{gridSearch0}
\alias{gridSearch0}
\title{gridSearch0}
\usage{
gridSearch0(LRjunc, SRjunc, thresSR = c(1:30), thresDis = c(1:20),
  adjustNCjunc = TRUE, matchedLS = NULL, fuzzyMatch = 100)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{thresSR}{}

\item{thresDis}{}

\item{adjustNCjunc}{}

\item{matchedLS}{}

\item{fuzzyMatch}{}
}
\description{
gridSearch0
}
