% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch2.R
\name{gridSearch2}
\alias{gridSearch2}
\title{gridSearch2}
\usage{
gridSearch2(LRjunc, SRjunc, thresSR = c(1:30), thresDis = c(1:20),
  adjustNCjunc = TRUE, matchedLS = NULL)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{thresSR}{}

\item{thresDis}{}

\item{adjustNCjunc}{}

\item{matchedLS}{}
}
\description{
gridSearch2
}
