% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch3.R
\name{gridSearch3}
\alias{gridSearch3}
\title{gridSearch3}
\usage{
gridSearch3(LRjunc, SRjunc, thresSR = c(1:30), thresDis = c(1:20),
  adjustNCjunc = TRUE, matchedLS = NULL)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{thresSR}{}

\item{thresDis}{}

\item{adjustNCjunc}{}

\item{matchedLS}{}
}
\description{
gridSearch3
}
