% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch4.R
\name{gridSearch4}
\alias{gridSearch4}
\title{gridSearch4}
\usage{
gridSearch4(LRjunc, SRjunc, thresSR = c(1:30), thresDis = c(1:20),
  adjustNCjunc = TRUE, matchedLS = NULL)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{thresSR}{}

\item{thresDis}{}

\item{adjustNCjunc}{}

\item{matchedLS}{}
}
\description{
gridSearch4
}
