% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juncToExon.R
\name{juncToExon}
\alias{juncToExon}
\title{juncToExon}
\usage{
juncToExon(juncList, s, e, exp = c())
}
\arguments{
\item{s}{}

\item{e}{}

\item{genome}{}

\item{chrom}{}
}
\description{
juncToExon transforms junc to exon
}
\examples{


}
