% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallistoQuant.R
\name{kallistoQuant}
\alias{kallistoQuant}
\title{kallistoQuant}
\usage{
kallistoQuant(transcript, SR1, SR2, outputDir)
}
\arguments{
\item{transcript}{}

\item{SR1}{}

\item{SR2}{}

\item{outputDir}{}
}
\description{
kallistoQuant
}
