% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGff.R
\name{matchGff}
\alias{matchGff}
\title{matchGff
finds the correspondence between two gffs}
\usage{
matchGff(gff1, gff2)
}
\arguments{
\item{gff1}{a list of data frame representing the junctions for each isoform}

\item{gff2}{a list of data frame representing the junctions for each isoform}
}
\value{
a vector of integer values indicating the index of gff2 matching gff1
}
\description{
matchGff
finds the correspondence between two gffs
}
\examples{


}
