% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLSjunc.R
\name{matchLSjunc}
\alias{matchLSjunc}
\title{matchLSjunc
match the junction site between long reads and short reads}
\usage{
matchLSjunc(LRjunc, SRjunc)
}
\arguments{
\item{LRjunc}{a list of data frame, indicating the splicing junction sites 
obtained from long reads}

\item{SRjunc}{a list of data frame, indicating the splicing junction sites 
obtained from short reads}
}
\value{
a list of data frame for the junction sites of each chromosome. The 
data frame indicates the locations of junction sites and the 
correspoding read count from both long and short reads
}
\description{
matchLSjunc
match the junction site between long reads and short reads
}
