% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLSjuncOneChr.R
\name{matchLSjuncOneChr}
\alias{matchLSjuncOneChr}
\title{matchLSjuncOneChr}
\usage{
matchLSjuncOneChr(lrc, src, fuzzyMatch = TRUE)
}
\arguments{
\item{lrc}{}

\item{src}{}

\item{fuzzyMatch}{}
}
\description{
matchLSjuncOneChr
}
\examples{

}
