% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchLSjuncOneChr1.R
\name{matchLSjuncOneChr1}
\alias{matchLSjuncOneChr1}
\title{matchLSjuncOneChr1}
\usage{
matchLSjuncOneChr1(lrc, src, fuzzyMatch = TRUE)
}
\arguments{
\item{lrc}{}

\item{src}{}

\item{fuzzyMatch}{}
}
\description{
matchLSjuncOneChr1
}
\examples{

}
