% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phqvExp.R
\name{phqvExp}
\alias{phqvExp}
\title{phqvExp}
\usage{
phqvExp(phqv, outputDir)
}
\arguments{
\item{phqv}{character value indicating the Isoseq polished high QV trascripts
in fasta/fastq, where}

\item{outputDir}{character value indicating the direcotry where results are 
saved.}
}
\value{
data frame representing the read count for each phqv consensus
}
\description{
phqvExp
}
