% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSmrtcell.R
\name{splitSmrtcell}
\alias{splitSmrtcell}
\title{splitSmrtcell}
\usage{
splitSmrtcell(alignments, outputDir, thres = 0)
}
\arguments{
\item{alignments}{}

\item{outputDir}{}

\item{thres}{}
}
\description{
splitSmrtcell
}
