% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR2bSMRT.R
\name{STAR2bSMRT}
\alias{STAR2bSMRT}
\title{STAR2bSMRT}
\usage{
STAR2bSMRT(genomeDir, genomeFasta, LR, SR1, SR2, thresSR, thresDis, outputDir,
  adjustNCjunc, fixedMatchedLS, chrom = NULL, s = 0, e = Inf, cores = 1)
}
\arguments{
\item{genomeDir}{}

\item{genomeFasta}{}

\item{LR}{}

\item{SR1}{}

\item{SR2}{}

\item{thresSR}{}

\item{thresDis}{}

\item{outputDir}{}

\item{adjustNCjunc}{}

\item{fixedMatchedLS}{}

\item{chrom}{}

\item{s}{}

\item{e}{}

\item{cores}{}
}
\description{
STAR2bSMRT
}
