% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR2bSMRT2.R
\name{STAR2bSMRT2}
\alias{STAR2bSMRT2}
\title{STAR2bSMRT2}
\usage{
STAR2bSMRT2(genomeDir, LR, SR1, SR2, outputDir, chrom = NULL, s = 0,
  e = Inf, cores)
}
\arguments{
\item{genomeDir}{}

\item{LR}{}

\item{SR1}{}

\item{SR2}{}

\item{outputDir}{}

\item{chrom}{}

\item{s}{}

\item{e}{}

\item{cores}{}
}
\description{
STAR2bSMRT2
}
