% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR2bSMRT_NRXN.R
\name{STAR2bSMRT_NRXN}
\alias{STAR2bSMRT_NRXN}
\title{STAR2bSMRT_NRXN}
\usage{
STAR2bSMRT_NRXN(genomeDir, genomeFasta, LRphqv = NULL, LRflnc = NULL,
  LRnfl = NULL, SR1, SR2 = NULL, useSJout = TRUE, adjustNCjunc = FALSE,
  thresSR, thresDis, outputDir, fixedMatchedLS = FALSE, fuzzyMatch = 100,
  chrom = NULL, s = 0, e = Inf, cores = 10, SoutputDir, LoutputDir)
}
\arguments{
\item{genomeDir}{}

\item{genomeFasta}{}

\item{SR1}{}

\item{SR2}{}

\item{useSJout}{}

\item{adjustNCjunc}{}

\item{thresSR}{}

\item{thresDis}{}

\item{outputDir}{}

\item{fixedMatchedLS}{}

\item{chrom}{}

\item{s}{}

\item{e}{}

\item{cores}{}

\item{phqv}{}

\item{flnc}{}

\item{nfl}{}
}
\description{
STAR2bSMRT_NRXN
}
\examples{

}
