% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starShort.R
\name{starShort}
\alias{starShort}
\title{starShort}
\usage{
starShort(genomeDir, SR1, SR2, outputDir)
}
\arguments{
\item{genomeDir}{}

\item{SR1}{}

\item{SR2}{}

\item{outputDir}{}
}
\description{
starShort
}
