% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionGff.R
\name{unionGff}
\alias{unionGff}
\title{unionGff
unions two isoform gff files}
\usage{
unionGff(gff1, gff2)
}
\arguments{
\item{gff1}{}

\item{gff2}{}
}
\description{
unionGff
unions two isoform gff files
}
\examples{


}
