% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagramGff.R
\name{vennDiagramGff}
\alias{vennDiagramGff}
\title{vennDiagramGff}
\usage{
vennDiagramGff(gffList, filename)
}
\arguments{
\item{gffList}{}

\item{filename}{}
}
\description{
vennDiagramGff
}
\examples{

}
