% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrectedIsoform.R
\name{generateCorrectedIsoform}
\alias{generateCorrectedIsoform}
\title{generateCorrectedIsoform}
\usage{
generateCorrectedIsoform(LRjunc, SRjunc, LRread, matchedLS, ts, td)
}
\arguments{
\item{LRjunc}{}

\item{SRjunc}{}

\item{LRread}{}

\item{matchedLS}{}

\item{ts}{}

\item{td}{}
}
\description{
generateCorrectedIsoform
}
