% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeq.R
\name{generateSeq}
\alias{generateSeq}
\title{generateSeq}
\usage{
generateSeq(genome, isoform, exp, chrom, s, e)
}
\arguments{
\item{genome}{}

\item{chrom}{}

\item{s}{}

\item{e}{}
}
\description{
generateSeq
}
