% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJuncBySJout.R
\name{getJuncBySJout}
\alias{getJuncBySJout}
\title{getJuncBySJout}
\usage{
getJuncBySJout(SJout = "SJ.out.tab", outputDir, chrom = NULL, s = 0,
  e = Inf)
}
\arguments{
\item{SJout}{a data frame, indicating the SJ.out.tab file ouput by STAR mapping}

\item{outputDir}{a character value, indicating the directory for saving results}

\item{chrom}{a character value, incidating the chromosome of interest}

\item{s}{an integer value indicating the start site of gene of interest. This
is useful for targeted sequencing}

\item{e}{an integer value indicating the end site of gene of interest. This
is useful for targeted sequencing}
}
\value{
a list of data frame representing the splicing junction sites for 
each chromosome
}
\description{
getJuncBySjout gets the unique junction sites and their numbers from the alignment files
}
\examples{

}
