% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRinfo.R
\name{getLRinfo}
\alias{getLRinfo}
\title{getLRinfo}
\usage{
getLRinfo(alignments, phqv = NULL, outputDir, chrom, s, e)
}
\arguments{
\item{alignments}{}

\item{phqv}{}

\item{outputDir}{}

\item{chrom}{}

\item{s}{}

\item{e}{}
}
\description{
getLRinfo
}
