% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRinfo.R
\name{getLRinfo}
\alias{getLRinfo}
\title{getLRinfo}
\usage{
getLRinfo(LRread, chrom = NULL, s = 0, e = Inf)
}
\arguments{
\item{LRread}{a data frame for the result of long read alignments, comprising
read id, start, end, junction sites}

\item{chrom}{}

\item{s}{}

\item{e}{}
}
\description{
getLRinfo
}
\examples{

exp = phqvExp( phqv , outputDir )
exp = exp$full_length_coverage
group = sapply( strsplit(as.character(LRread$id),"/"), function(x) x[1] )

}
