% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallistoQuant.R
\name{kallistoQuant}
\alias{kallistoQuant}
\title{kallistoQuant
perform the kallisto transcript quantification}
\usage{
kallistoQuant(transcript, SR1, SR2, outputDir)
}
\arguments{
\item{transcript}{character value indicating the transcriptome file}

\item{SR1}{character value indicating the short read file in fastq format: 
single-end or paired-end R1}

\item{SR2}{character value indicating the short read file in fastq format: 
paired-end R2}

\item{outputDir}{character value indicating the direcotry where results are 
saved.}
}
\description{
kallistoQuant
perform the kallisto transcript quantification
}
