% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonQuant.R
\name{salmonQuant}
\alias{salmonQuant}
\title{salmonQuant}
\usage{
salmonQuant(transcript, SR1, SR2, outputDir)
}
\arguments{
\item{transcript}{}

\item{SR1}{}

\item{SR2}{}

\item{outputDir}{}
}
\description{
salmonQuant
}
