% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR2bSMRT.R
\name{STAR2bSMRT}
\alias{STAR2bSMRT}
\title{STAR2bSMRT}
\usage{
STAR2bSMRT(genomeDir, genomeFasta, phqv = NULL, flnc = NULL, nfl = NULL,
  SR1, SR2 = NULL, useSJout = TRUE, adjustNCjunc = FALSE, thresSR,
  thresDis, outputDir, fixedMatchedLS = FALSE, chrom = NULL, s = 0,
  e = Inf, cores = 1)
}
\arguments{
\item{genomeDir}{}

\item{genomeFasta}{}

\item{phqv}{}

\item{flnc}{}

\item{nfl}{}

\item{SR1}{}

\item{SR2}{}

\item{useSJout}{}

\item{adjustNCjunc}{}

\item{thresSR}{}

\item{thresDis}{}

\item{outputDir}{}

\item{fixedMatchedLS}{}

\item{chrom}{}

\item{s}{}

\item{e}{}

\item{cores}{}
}
\description{
STAR2bSMRT
}
