% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starLong.R
\name{starLong}
\alias{starLong}
\title{starLong}
\usage{
starLong(genomeDir, LR, outputDir, cores = 10, SJ = NULL)
}
\arguments{
\item{genomeDir}{}

\item{LR}{}

\item{outputDir}{}

\item{cores}{}

\item{SJ}{}
}
\description{
starLong
}
