% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionGff.R
\name{unionGff}
\alias{unionGff}
\title{unionGff
unions two isoform gff files}
\usage{
unionGff(gff1, gff2)
}
\arguments{
\item{gff1}{a list of data frame representing the junctions for each isoform}

\item{gff2}{a list of data frame representing the junctions for each isoform}
}
\value{
a combined list of unique isoforms
}
\description{
unionGff
unions two isoform gff files
}
\examples{


}
