% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennDiagramGff.R
\name{vennDiagramGff}
\alias{vennDiagramGff}
\title{vennDiagramGff
draw the venn diagram of a list of gffs}
\usage{
vennDiagramGff(gffList, filename)
}
\arguments{
\item{gffList}{a list of gffs representing gffs for different samples}

\item{filename}{character value representing the file name of VennDiagram}
}
\description{
vennDiagramGff
draw the venn diagram of a list of gffs
}
\examples{

}
