% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGff.R
\name{writeGff}
\alias{writeGff}
\title{writeGff}
\usage{
writeGff(isoform, file = "", exp, chrom, s, e)
}
\arguments{
\item{isoform}{}

\item{file}{}

\item{exp}{}

\item{chrom}{}

\item{s}{}

\item{e}{}
}
\description{
writeGff
}
\examples{


}
